tests/wpt/web-platform-tests/css/css-lists/list-and-block-textarea-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Lists: test list with textarea as its first child</title>
<link rel=help href="https://www.w3.org/TR/CSS22/generate.html#lists">
<!-- https://bugs.chromium.org/p/chromium/issues/detail?id=767408 -->

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
textarea {
  border: 0px;
  padding: 0px;
}
</style>

<div id="log"></div>

<ul>
  <li id="target">
    <textarea rows="3" cols="20" style="display:block; height:45px">
      hello
    </textarea>
  </li>
</ul>

<script>
test(function() {
  var height = document.getElementById("target").offsetHeight;
  assert_equals(height, 45, "the height of li should be 45px, and no extra line generated")
}, "list and block textarea");

</script>
